refactor: MoreOptionsIconContainer
refactor into own component and reuse
